+++ src/app/(landing)/subsidiariesandinvesment/item/[id]/page.tsx
