+++ src/app/api/subsidiariesandinvesment-category/[id]/route.ts
