+++ src/app/api/subsidiariesandinvesment-item/[id]/route.ts
